rename resources_util.js -> resources_api_util.js
